# Allow path params with conflicting names and different methods