Closes issues #2201 and #1726

What was done:
1. Each node is now splitted into per-method handler with it's own ppath and pnames
2. Test case from #1726 at [router_test.go](https://github.com/labstack/echo/pull/2207/files#diff-06b5e9693dbcea8c23abd4249e8b42aab7cb9affcc5366568e6a34b212c62ff5R824-R845)