@ortyomka 2022-06-29
Sorry, but you copy my solution from #2208. I think, it's not good.

Proof (force-push): https://github.com/labstack/echo/compare/81ceb2f7e4a38689693dadf79c5ce4c5ca2f46b4..459e309d2067f032d0984f05d718ad4b13b90b8c