@aldas 2022-06-29
I'll check it soon but there seems to be little bit too few tests for huge change like that. 

I see 3 things here:
1. performance: route is searched by method - thus more uncommon method types are routed faster as trees are smaller
2. functionality: we (could) support any method type instead of `v4` fixed set. note: in `v5` branch we have similar feature added but instead of adding all method I used additional map with current methof fields.
3. user experience: each route has their own instance of path parameter names. note: in `v5` branch  we have similar feature added.  
4. 
5. I'm little but suspicious if this solution is not a little bit too naive. Having tree for each method and making sure that correct allowheader is composed is possibly a lot more complex due now having to find matching methods from multiple trees and what implications this has too OPTIONS performance