@dmitrybarsukov 2022-06-29
@ortyomka hello!
Yes, i've force-pushed new solution because my previous solution (separate trees for each tree node) had performance issues and much more allocs/op than original router.
But it's not copy of your solution.