@dmitrybarsukov 2022-06-29
@aldas 
Sorry, my solution has serious performance issues.

#### Original benchmark
```
BenchmarkEchoStaticRoutes
BenchmarkEchoStaticRoutes-8                    	  113509	     10525 ns/op	       0 B/op	       0 allocs/op
BenchmarkEchoStaticRoutesMisses
BenchmarkEchoStaticRoutesMisses-8              	  113067	     10611 ns/op	       0 B/op	       0 allocs/op
BenchmarkEchoGitHubAPI
BenchmarkEchoGitHubAPI-8                       	   56080	     21848 ns/op	       0 B/op	       0 allocs/op
BenchmarkEchoGitHubAPIMisses
BenchmarkEchoGitHubAPIMisses-8                 	   57042	     21749 ns/op	       0 B/op	       0 allocs/op
BenchmarkEchoParseAPI
BenchmarkEchoParseAPI-8                        	  913242	      1270 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterStaticRoutes
BenchmarkRouterStaticRoutes-8                  	  149326	      8167 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterStaticRoutesMisses
BenchmarkRouterStaticRoutesMisses-8            	 3681016	       300.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGitHubAPI
BenchmarkRouterGitHubAPI-8                     	   75144	     15802 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGitHubAPIMisses
BenchmarkRouterGitHubAPIMisses-8               	 3314385	       361.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterParseAPI
BenchmarkRouterParseAPI-8                      	 1495852	       795.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterParseAPIMisses
BenchmarkRouterParseAPIMisses-8                	 6146856	       193.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGooglePlusAPI
BenchmarkRouterGooglePlusAPI-8                 	 2260587	       531.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGooglePlusAPIMisses
BenchmarkRouterGooglePlusAPIMisses-8           	 4245891	       282.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterParamsAndAnyAPI
BenchmarkRouterParamsAndAnyAPI-8               	  901524	      1293 ns/op	       0 B/op	       0 allocs/op
```

#### map of method to tree solution
```
BenchmarkEchoStaticRoutes
BenchmarkEchoStaticRoutes-8                    	  102147	     11708 ns/op	       0 B/op	       0 allocs/op
BenchmarkEchoStaticRoutesMisses
BenchmarkEchoStaticRoutesMisses-8              	  102079	     11731 ns/op	       0 B/op	       0 allocs/op
BenchmarkEchoGitHubAPI
BenchmarkEchoGitHubAPI-8                       	   39241	     30526 ns/op	   15108 B/op	     236 allocs/op
BenchmarkEchoGitHubAPIMisses
BenchmarkEchoGitHubAPIMisses-8                 	   38858	     30129 ns/op	   15108 B/op	     236 allocs/op
BenchmarkEchoParseAPI
BenchmarkEchoParseAPI-8                        	  509515	      2271 ns/op	     832 B/op	      26 allocs/op
BenchmarkRouterStaticRoutes
BenchmarkRouterStaticRoutes-8                  	  125223	      9895 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterStaticRoutesMisses
BenchmarkRouterStaticRoutesMisses-8            	 1000000	      1195 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGitHubAPI
BenchmarkRouterGitHubAPI-8                     	   45582	     26628 ns/op	   15108 B/op	     236 allocs/op
BenchmarkRouterGitHubAPIMisses
BenchmarkRouterGitHubAPIMisses-8               	  381775	      3166 ns/op	    1024 B/op	      16 allocs/op
BenchmarkRouterParseAPI
BenchmarkRouterParseAPI-8                      	  603487	      2012 ns/op	     832 B/op	      26 allocs/op
BenchmarkRouterParseAPIMisses
BenchmarkRouterParseAPIMisses-8                	  555308	      2163 ns/op	     512 B/op	      16 allocs/op
BenchmarkRouterGooglePlusAPI
BenchmarkRouterGooglePlusAPI-8                 	 1000000	      1067 ns/op	     416 B/op	      13 allocs/op
BenchmarkRouterGooglePlusAPIMisses
BenchmarkRouterGooglePlusAPIMisses-8           	  565418	      2119 ns/op	     512 B/op	      16 allocs/op
BenchmarkRouterParamsAndAnyAPI
BenchmarkRouterParamsAndAnyAPI-8               	  476144	      2556 ns/op	    1920 B/op	      24 allocs/op
``` 
As you see, it has much more ns/op and >0 allocs/op.
So, i had to rewrite router logic and make map of method to handler inside each node.
This variant gives much better results.
```
BenchmarkEchoStaticRoutes
BenchmarkEchoStaticRoutes-8                    	   93038	     13065 ns/op	       0 B/op	       0 allocs/op
BenchmarkEchoStaticRoutesMisses
BenchmarkEchoStaticRoutesMisses-8              	   91372	     13125 ns/op	       0 B/op	       0 allocs/op
BenchmarkEchoGitHubAPI
BenchmarkEchoGitHubAPI-8                       	   47766	     26254 ns/op	       0 B/op	       0 allocs/op
BenchmarkEchoGitHubAPIMisses
BenchmarkEchoGitHubAPIMisses-8                 	   47211	     25293 ns/op	       0 B/op	       0 allocs/op
BenchmarkEchoParseAPI
BenchmarkEchoParseAPI-8                        	  673542	      1790 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterStaticRoutes
BenchmarkRouterStaticRoutes-8                  	  105259	     11410 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterStaticRoutesMisses
BenchmarkRouterStaticRoutesMisses-8            	 3792248	       312.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGitHubAPI
BenchmarkRouterGitHubAPI-8                     	   56343	     21262 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGitHubAPIMisses
BenchmarkRouterGitHubAPIMisses-8               	 3242745	       370.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterParseAPI
BenchmarkRouterParseAPI-8                      	  907677	      1313 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterParseAPIMisses
BenchmarkRouterParseAPIMisses-8                	 5698016	       211.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGooglePlusAPI
BenchmarkRouterGooglePlusAPI-8                 	 1675465	       716.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGooglePlusAPIMisses
BenchmarkRouterGooglePlusAPIMisses-8           	 3999283	       302.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterParamsAndAnyAPI
BenchmarkRouterParamsAndAnyAPI-8               	  578578	      2063 ns/op	       0 B/op	       0 allocs/op
```

Can you please re-review my PR again?